Changed uint8 to float32 for input image.

+++ research/delf/delf/python/training/model/export_local_and_global_model.py
@@ -1,3 +1,12 @@
+# @Author: Jan Brejcha <janbrejcha>
+# @Date:   2020-11-18T16:31:04+01:00
+# @Email:  ibrejcha@fit.vutbr.cz, brejchaja@gmail.com
+# @Project: Locate
+# @Last modified by:   janbrejcha
+# @Last modified time: 2020-12-10T13:00:13+01:00
+
+
+
 # Lint as: python3
 # Copyright 2020 The TensorFlow Authors. All Rights Reserved.
 #
@@ -94,7 +103,7 @@ class _ExtractModule(tf.Module):
     self._model.load_weights(checkpoint_path)
 
   @tf.function(input_signature=[
-      tf.TensorSpec(shape=[None, None, 3], dtype=tf.uint8, name='input_image'),
+      tf.TensorSpec(shape=[None, None, 3], dtype=tf.float32, name='input_image'),
       tf.TensorSpec(shape=[None], dtype=tf.float32, name='input_scales'),
       tf.TensorSpec(shape=(), dtype=tf.int32, name='input_max_feature_num'),
       tf.TensorSpec(shape=(), dtype=tf.float32, name='input_abs_thres'),
